LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-fonts/cjk-kerning-expected.txt

Testing application of kerning to CJK text

default
YeYeYeYeティティティティ
font-kerning: none;
YeYeYeYeティティティティ
font-kerning: normal;
YeYeYeYeティティティティ
font-feature-settings:"palt" 0; font-kerning: normal;
YeYeYeYeティティティティ
font-feature-settings:"palt" 1; font-kerning: normal;
YeYeYeYeティティティティ
font-feature-settings:"palt" 1; font-kerning: none;
YeYeYeYeティティティティ

PASS expected match: .kernON .latin vs .paltOFFkernON .latin
PASS expected match: .kernON .latin vs .paltONkernON .latin
PASS expected match: .kernOFF .latin vs .paltONkernOFF .latin
PASS expected match: .kernON .cjk vs .paltONkernON .cjk
FAIL expected match: .default .cjk vs .kernOFF .cjk assert_equals: expected 384 but got 385
PASS expected mismatch: .kernOFF .latin vs .kernON .latin
PASS expected mismatch: .kernOFF .cjk vs .kernON .cjk
FAIL expected mismatch: .paltOFFkernON .cjk vs .paltONkernON .cjk assert_not_equals: got disallowed value 385
PASS .default .latin matches one of [.kernON .latin, .kernOFF .latin]
